requirements.txt
certifi==2019.11.28
dj-database-url==0.5.0
Django==2.2.5
django-versatileimagefield==2.0
django-phone-field==1.8.0
Pillow==7.0.0
psycopg2==2.7.4
pytz==2019.3
sqlparse==0.3.1
django-prices==2.2.0
